Moved as_mutable() and as_immutable() to MatrixBase
